requirements.txt
git+https://github.com/androguard/androguard.git@v3.1.0-pre.2
requests==2.18.4
Pillow==5.0.0
dhash==1.3
jellyfish==0.5.6